Move oracle fetcher config to sdk
