+++ src/lib/oracleKeeperFetcher/oracleKeeperFetcher.ts
@@ -1,5 +1,5 @@
 import { isLocal } from "config/env";
-import { getOracleKeeperNextIndex, getOracleKeeperUrl } from "config/oracleKeeper";
+import { getOracleKeeperNextIndex, getOracleKeeperUrl } from "sdk/configs/oracleKeeper";
 import { getNormalizedTokenSymbol } from "sdk/configs/tokens";
 import { Bar, FromNewToOldArray } from "domain/tradingview/types";
 import { buildUrl } from "sdk/utils/buildUrl";
